[refactor] 예외 상속하도록 리팩토링 (#17)
