+++ src/main/java/org/recordy/server/auth/exception/AuthException.java
@@ -1,14 +1,11 @@
 package org.recordy.server.auth.exception;
 
-import org.recordy.server.auth.message.ErrorMessage;
+import org.recordy.server.common.message.ErrorMessage;
+import org.recordy.server.common.exception.RecordyException;
 
-public class AuthException extends RuntimeException {
-
-    public AuthException() {
-        super(ErrorMessage.UNAUTHORIZED.getMessage());
-    }
+public class AuthException extends RecordyException {
 
     public AuthException(ErrorMessage errorMessage) {
-        super(errorMessage.getMessage());
+        super(errorMessage);
     }
 }
