--- src/main/java/org/recordy/server/auth/message/ErrorMessage.java
@@ -1,35 +1,0 @@
-package org.recordy.server.auth.message;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public enum ErrorMessage {
-
-    /**
-     * 401 Unauthorized
-     */
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "리소스 접근 권한이 없습니다."),
-    INVALID_KAKAO_ACCESS_TOKEN(HttpStatus.UNAUTHORIZED, "카카오 액세스 토큰의 정보를 조회하는 과정에서 오류가 발생하였습니다."),
-    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "액세스 토큰의 형식이 올바르지 않습니다. Bearer 타입을 확인해 주세요."),
-    INVALID_TOKEN_VALUE(HttpStatus.UNAUTHORIZED, "액세스 토큰의 값이 올바르지 않습니다."),
-    EXPIRED_TOKEN(HttpStatus.UNAUTHORIZED, "액세스 토큰이 만료되었습니다. 재발급 받아주세요."),
-    NOT_MATCH_REFRESH_TOKEN(HttpStatus.UNAUTHORIZED, "일치하지 않는 리프레시 토큰입니다."),
-
-    /**
-     * 404 Not Found
-     */
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "존재하지 않는 회원입니다."),
-
-    /**
-     * 409 Conflict
-     */
-    DUPLICATE_NICKNAME(HttpStatus.CONFLICT, "이미 존재하는 닉네임입니다."),
-    DUPLICATE_USER(HttpStatus.CONFLICT, "이미 존재하는 회원입니다.");
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
