+++ src/main/java/org/recordy/server/auth/service/impl/kakao/AuthKakaoPlatformServiceImpl.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.recordy.server.auth.domain.AuthPlatform;
 import org.recordy.server.auth.domain.usecase.AuthSignIn;
 import org.recordy.server.auth.exception.AuthException;
-import org.recordy.server.auth.message.ErrorMessage;
 import org.recordy.server.auth.service.AuthPlatformService;
+import org.recordy.server.common.message.ErrorMessage;
 import org.springframework.stereotype.Service;
 
 @Slf4j
