+++ src/main/java/org/recordy/server/common/exception/RecordyException.java
@@ -5,11 +5,11 @@ import org.recordy.server.common.message.ErrorMessage;
 
 
 @Getter
-public class CommonException extends RuntimeException {
+public class RecordyException extends RuntimeException {
 
     private final ErrorMessage errorMessage;
 
-    public CommonException(ErrorMessage errorMessage) {
+    public RecordyException(ErrorMessage errorMessage) {
         super(errorMessage.getMessage());
         this.errorMessage = errorMessage;
     }
